refactor(config): move module Config.load calls to config files
