+++ src/database/index.ts
@@ -1,3 +1,5 @@
+import './config';
+
 export * from './database.module';
 export * from './database.service';
 export * from './database.types';
