+++ src/http-core/http-core.types.ts
@@ -21,20 +21,6 @@ declare module 'fastify' {
   }
 }
 
-declare module '@shadow-library/common' {
-  export interface ConfigRecords {
-    'http-core.csrf.enabled': boolean;
-    'http-core.helmet.enabled'?: boolean;
-    'http-core.compress.enabled'?: boolean;
-    'http-core.openapi.enabled'?: boolean;
-
-    /*! Health Service configs */
-    'health.host': string;
-    'health.port': number;
-    'health.enabled': boolean;
-  }
-}
-
 export interface CommonOptions {
   enabled?: boolean;
 }
