+++ src/http-core/index.ts
@@ -1,3 +1,5 @@
+import './config';
+
 export * from './dtos';
 export * from './http-core.module';
 export * from './services';
